+++ blueprints/__init__.py
@@ -2,5 +2,5 @@
 
 default_app_config = "blueprints.apps.BlueprintsConfig"
 
-__version__ = "1.5.2"
+__version__ = "1.6.0"
 __title__ = "Blueprints"