+++ blueprints/admin.py
@@ -3,12 +3,13 @@
 from typing import Any
 
 from django.contrib import admin
-from django.db.models.query import QuerySet
+from django.db.models import QuerySet
 from django.http.request import HttpRequest
 from django.urls import reverse
 from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
+from . import tasks
 from .models import Blueprint, IndustryJob, Location, Owner, Request
 
 
@@ -52,7 +53,7 @@ class LocationFlagListFilter(admin.SimpleListFilter):
     def lookups(self, request, model_admin):
         qs = model_admin.get_queryset(request)
         return (
-            (obj, obj)
+            (obj, Blueprint.LocationFlag(obj).label)
             for obj in qs.values_list("location_flag", flat=True)
             .distinct()
             .order_by("location_flag")
@@ -122,6 +123,36 @@ class BlueprintAdmin(admin.ModelAdmin):
         return False
 
 
+class LocationHasBlueprintsListFilter(admin.SimpleListFilter):
+    title = _("has blueprints")
+    parameter_name = "has_blueprints"
+
+    def lookups(self, request, model_admin):
+        return (("yes", _("yes")), ("no", _("no")))
+
+    def queryset(self, request, queryset):
+        value = self.value()
+        if value == "yes":
+            return queryset.annotate_blueprint_count().filter(blueprint_count__gt=0)
+        if value == "no":
+            return queryset.annotate_blueprint_count().filter(blueprint_count=0)
+
+
+class LocationHasNameListFilter(admin.SimpleListFilter):
+    title = _("has name")
+    parameter_name = "has_name"
+
+    def lookups(self, request, model_admin):
+        return (("yes", _("yes")), ("no", _("no")))
+
+    def queryset(self, request, queryset):
+        value = self.value()
+        if value == "yes":
+            return queryset.exclude(name="")
+        if value == "no":
+            return queryset.filter(name="")
+
+
 @admin.register(Location)
 class LocationAdmin(admin.ModelAdmin):
     list_display = (
@@ -131,10 +162,13 @@ class LocationAdmin(admin.ModelAdmin):
         "_group",
         "_solar_system",
         "_parent",
+        "_blueprint_count",
         "updated_at",
     )
     list_filter = (
         # ("parent", admin.RelatedOnlyFieldListFilter),
+        LocationHasBlueprintsListFilter,
+        LocationHasNameListFilter,
         (
             "eve_solar_system__eve_constellation__eve_region",
             admin.RelatedOnlyFieldListFilter,
@@ -153,7 +187,13 @@ class LocationAdmin(admin.ModelAdmin):
             "eve_type__eve_group",
             "parent",
             "parent__eve_type",
-        )
+        ).annotate_blueprint_count()
+
+    def has_add_permission(self, request):
+        return False
+
+    def has_change_permission(self, request, obj=None):
+        return False
 
     @admin.display(ordering="eve_solar_system__name")
     def _solar_system(self, obj: Location):
@@ -173,25 +213,52 @@ class LocationAdmin(admin.ModelAdmin):
         url = reverse("admin:blueprints_location_change", args=(obj.parent.id,))
         return format_html('<a href="{}">{}</a>', url, obj.parent)
 
-    def has_add_permission(self, request):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False
+    @admin.display(ordering="blueprint_count")
+    def _blueprint_count(self, obj: Location):
+        return obj.blueprint_count
 
 
 @admin.register(Owner)
 class OwnerAdmin(admin.ModelAdmin):
-    list_display = ("character", "_type", "corporation", "is_active")
+    list_display = ("__str__", "is_active", "_type", "character", "_blueprint_count")
+    actions = ["activate_owners", "deactivate_owners", "update_locations"]
 
-    def _type(self, obj):
-        return "Corporate" if obj.corporation else "Personal"
+    def get_queryset(self, request: HttpRequest) -> QuerySet[Any]:
+        qs = (
+            super()
+            .get_queryset(request)
+            .select_related("character__character", "corporation", "character__user")
+            .annotate_blueprint_count()
+        )
+        return qs
 
     def has_add_permission(self, request):
         return False
 
     def has_change_permission(self, request, obj=None):
         return False
+
+    @admin.display(ordering="blueprint_count")
+    def _blueprint_count(self, obj: Owner):
+        return obj.blueprint_count
+
+    def _type(self, obj):
+        return "Corporate" if obj.corporation else "Personal"
+
+    @admin.action(description="Update locations for selected owners")
+    def update_locations(self, request, queryset):
+        for owner in queryset:
+            tasks.update_locations_for_owner.delay(owner_pk=owner.pk)
+        count = queryset.count()
+        self.message_user(request, f"Started updating locations for {count} owners.")
+
+    @admin.action(description="Activate selected owners")
+    def activate_owners(self, request, queryset):
+        queryset.update(is_active=True)
+
+    @admin.action(description="Deactivate selected owners")
+    def deactivate_owners(self, request, queryset):
+        queryset.update(is_active=False)
 
 
 @admin.register(Request)