+++ blueprints/models.py
@@ -1,11 +1,11 @@
 """Models for Blueprints."""
 
-from typing import List, Tuple
+from typing import List
 
 from django.contrib.auth.models import Permission, User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-from esi.errors import TokenExpiredError, TokenInvalidError
+from esi.errors import TokenError
 from esi.models import Token
 from eveuniverse.models import EveEntity, EveSolarSystem, EveType
 
@@ -15,18 +15,23 @@ from allianceauth.eveonline.models import EveCharacter, EveCorporationInfo
 from allianceauth.notifications import notify
 from allianceauth.services.hooks import get_extension_logger
 from app_utils.django import users_with_permission
-from app_utils.logging import LoggerAddTag, make_logger_prefix
+from app_utils.logging import LoggerAddTag
 
 from . import __title__
-from .constants import EVE_LOCATION_FLAGS
 from .decorators import fetch_token_for_owner
-from .managers import BlueprintManager, LocationManager, RequestManager
+from .managers import BlueprintManager, LocationManager, OwnerManager, RequestManager
 from .providers import esi
 from .validators import validate_material_efficiency, validate_time_efficiency
 
 NAMES_MAX_LENGTH = 100
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
+
+
+def get_or_create_location_async(location_id: int, token: Token) -> "Location":
+    """Get or create a location sync - helper function."""
+    obj, _ = Location.objects.get_or_create_esi_async(id=location_id, token=token)
+    return obj
 
 
 class General(models.Model):
@@ -49,29 +54,6 @@ class General(models.Model):
 
 class Owner(models.Model):
     """A corporation that owns blueprints"""
-
-    ERROR_NONE = 0
-    ERROR_TOKEN_INVALID = 1
-    ERROR_TOKEN_EXPIRED = 2
-    ERROR_INSUFFICIENT_PERMISSIONS = 3
-    ERROR_NO_CHARACTER = 4
-    ERROR_ESI_UNAVAILABLE = 5
-    ERROR_OPERATION_MODE_MISMATCH = 6
-    ERROR_UNKNOWN = 99
-
-    ERRORS_LIST = [
-        (ERROR_NONE, "No error"),
-        (ERROR_TOKEN_INVALID, "Invalid token"),
-        (ERROR_TOKEN_EXPIRED, "Expired token"),
-        (ERROR_INSUFFICIENT_PERMISSIONS, "Insufficient permissions"),
-        (ERROR_NO_CHARACTER, "No character set for fetching data from ESI"),
-        (ERROR_ESI_UNAVAILABLE, "ESI API is currently unavailable"),
-        (
-            ERROR_OPERATION_MODE_MISMATCH,
-            "Operaton mode does not match with current setting",
-        ),
-        (ERROR_UNKNOWN, "Unknown error"),
-    ]
 
     corporation = models.OneToOneField(
         EveCorporationInfo,
@@ -96,6 +78,8 @@ class Owner(models.Model):
         help_text=("whether this owner is currently included in the sync process"),
     )
 
+    objects = OwnerManager()
+
     class Meta:
         default_permissions = ()
 
@@ -107,24 +91,36 @@ class Owner(models.Model):
         try:
             if self.corporation:
                 return self.corporation.corporation_name
-            return self.character.character.character_name
-        except AttributeError:
+            return self.eve_character_strict.character_name
+        except (ValueError, AttributeError):
             return ""
+
+    @property
+    def corporation_strict(self) -> EveCorporationInfo:
+        if not self.corporation:
+            raise ValueError("No corporation defined")
+        return self.corporation
+
+    @property
+    def eve_character_strict(self) -> EveCharacter:
+        if not self.character or not self.character.character:
+            raise ValueError("No character defined")
+        return self.character.character
 
     def update_locations_esi(self):
         if self.corporation:
             assets = self._fetch_corporate_assets()
-            token = self.token(
+            token = self.valid_token(
                 [
                     "esi-universe.read_structures.v1",
                     "esi-assets.read_corporation_assets.v1",
                 ]
-            )[0]
+            )
         else:
             assets = self._fetch_personal_assets()
-            token = self.token(
+            token = self.valid_token(
                 ["esi-universe.read_structures.v1", "esi-assets.read_assets.v1"]
-            )[0]
+            )
 
         asset_ids = []
         asset_locations = {}
@@ -141,93 +137,80 @@ class Owner(models.Model):
                 else:
                     asset_locations[location_id] = [asset["item_id"]]
 
-        for location in list(asset_locations.keys()):
-            if not Location.objects.filter(id=location).count() > 0:
-                parent_location = assets_by_id[location]["location_id"]
-                Location.objects.create(
-                    id=location,
-                    parent=self._fetch_location(parent_location, token=token),
-                    eve_type=EveType.objects.filter(
-                        id=assets_by_id[location]["type_id"]
-                    ).first(),
-                )
-            else:
-                parent_location = assets_by_id[location]["location_id"]
-                location_obj = Location.objects.filter(id=location).first()
-                if location_obj:
-                    location_obj.parent = self._fetch_location(
-                        parent_location, token=token
-                    )
-                    location_obj.save()
+        for location_id in asset_locations.keys():
+            asset = assets_by_id[location_id]
+            parent_location = asset["location_id"]
+            parent = get_or_create_location_async(parent_location, token=token)
+            eve_type, _ = EveType.objects.get_or_create_esi(id=asset["type_id"])
+            Location.objects.update_or_create(
+                id=location_id, defaults={"parent": parent, "eve_type": eve_type}
+            )
 
     def update_blueprints_esi(self):
-        """updates all blueprints from ESI"""
+        """Update all blueprints from ESI."""
 
-        if self.is_active:
-            blueprint_ids_to_remove = list(
-                Blueprint.objects.filter(owner=self).values_list("item_id", flat=True)
+        blueprint_ids_to_remove = list(
+            self.blueprints.values_list("item_id", flat=True)
+        )
+        if self.corporation:
+            blueprints = self._fetch_corporate_blueprints()
+            token = self.valid_token(
+                [
+                    "esi-universe.read_structures.v1",
+                    "esi-corporations.read_blueprints.v1",
+                ]
             )
-            if self.corporation:
-                blueprints = self._fetch_corporate_blueprints()
-                token = self.token(
-                    [
-                        "esi-universe.read_structures.v1",
-                        "esi-corporations.read_blueprints.v1",
-                    ]
-                )[0]
-            else:
-                blueprints = self._fetch_personal_blueprints()
-                token = self.token(
-                    [
-                        "esi-universe.read_structures.v1",
-                        "esi-characters.read_blueprints.v1",
-                    ]
-                )[0]
+        else:
+            blueprints = self._fetch_personal_blueprints()
+            token = self.valid_token(
+                [
+                    "esi-universe.read_structures.v1",
+                    "esi-characters.read_blueprints.v1",
+                ]
+            )
 
-            for blueprint in blueprints:
-                runs = blueprint["runs"]
-                if runs < 1:
-                    runs = None
-                quantity = blueprint["quantity"]
-                if quantity < 0:
-                    quantity = 1
-                original = Blueprint.objects.filter(
-                    owner=self, item_id=blueprint["item_id"]
-                ).first()
-                if original is not None:
-                    # We've seen this blueprint coming from ESI, so we know it shouldn't be deleted
-                    blueprint_ids_to_remove.remove(original.item_id)
-                    original.location = self._fetch_location(
-                        blueprint["location_id"],
-                        token=token,
-                    )
-                    original.location_flag = blueprint["location_flag"]
-                    original.eve_type = EveType.objects.get_or_create_esi(
-                        id=blueprint["type_id"]
-                    )[0]
-                    original.runs = runs
-                    original.material_efficiency = blueprint["material_efficiency"]
-                    original.time_efficiency = blueprint["time_efficiency"]
-                    original.quantity = quantity
-                    original.save()
-                else:
-                    Blueprint.objects.create(
-                        owner=self,
-                        location=self._fetch_location(
-                            blueprint["location_id"],
-                            token=token,
-                        ),
-                        location_flag=blueprint["location_flag"],
-                        eve_type=EveType.objects.get_or_create_esi(
-                            id=blueprint["type_id"]
-                        )[0],
-                        item_id=blueprint["item_id"],
-                        runs=runs,
-                        material_efficiency=blueprint["material_efficiency"],
-                        time_efficiency=blueprint["time_efficiency"],
-                        quantity=quantity,
-                    )
-            Blueprint.objects.filter(pk__in=blueprint_ids_to_remove).delete()
+        for blueprint in blueprints:
+            runs = blueprint["runs"]
+            if runs < 1:
+                runs = None
+            quantity = blueprint["quantity"]
+            if quantity < 0:
+                quantity = 1
+            original = self.blueprints.filter(item_id=blueprint["item_id"]).first()
+
+            location_flag = Blueprint.LocationFlag.from_esi_data(
+                blueprint["location_flag"]
+            )
+            eve_type, _ = EveType.objects.get_or_create_esi(id=blueprint["type_id"])
+            if original is not None:
+                # We've seen this blueprint coming from ESI, so we know it shouldn't be deleted
+                blueprint_ids_to_remove.remove(original.item_id)
+                original.location = get_or_create_location_async(
+                    blueprint["location_id"],
+                    token=token,
+                )
+                original.location_flag = location_flag
+                original.eve_type = eve_type
+                original.runs = runs
+                original.material_efficiency = blueprint["material_efficiency"]
+                original.time_efficiency = blueprint["time_efficiency"]
+                original.quantity = quantity
+                original.save()
+            else:
+                self.blueprints.create(
+                    location=get_or_create_location_async(
+                        blueprint["location_id"], token=token
+                    ),
+                    location_flag=location_flag,
+                    eve_type=eve_type,
+                    item_id=blueprint["item_id"],
+                    runs=runs,
+                    material_efficiency=blueprint["material_efficiency"],
+                    time_efficiency=blueprint["time_efficiency"],
+                    quantity=quantity,
+                )
+
+        Blueprint.objects.filter(pk__in=blueprint_ids_to_remove).delete()
 
     def update_industry_jobs_esi(self):
         """updates all blueprints from ESI"""
@@ -238,20 +221,20 @@ class Owner(models.Model):
             )
             if self.corporation:
                 jobs = self._fetch_corporate_industry_jobs()
-                token = self.token(
+                token = self.valid_token(
                     [
                         "esi-universe.read_structures.v1",
                         "esi-industry.read_corporation_jobs.v1",
                     ]
-                )[0]
+                )
             else:
                 jobs = self._fetch_personal_industry_jobs()
-                token = self.token(
+                token = self.valid_token(
                     [
                         "esi-universe.read_structures.v1",
                         "esi-industry.read_character_jobs.v1",
                     ]
-                )[0]
+                )
 
             for job in jobs:
                 original = IndustryJob.objects.filter(
@@ -278,7 +261,7 @@ class Owner(models.Model):
                                 id=job["job_id"],
                                 activity=job["activity_id"],
                                 owner=self,
-                                location=self._fetch_location(
+                                location=get_or_create_location_async(
                                     job["output_location_id"],
                                     token=token,
                                 ),
@@ -297,151 +280,238 @@ class Owner(models.Model):
     @fetch_token_for_owner(["esi-assets.read_corporation_assets.v1"])
     def _fetch_corporate_assets(self, token) -> list:
         return esi.client.Assets.get_corporations_corporation_id_assets(
-            corporation_id=self.corporation.corporation_id,
+            corporation_id=self.corporation_strict.corporation_id,
             token=token.valid_access_token(),
         ).results()
 
     @fetch_token_for_owner(["esi-assets.read_assets.v1"])
     def _fetch_personal_assets(self, token) -> list:
         return esi.client.Assets.get_characters_character_id_assets(
-            character_id=self.character.character.character_id,
+            character_id=self.eve_character_strict.character_id,
             token=token.valid_access_token(),
         ).results()
 
     @fetch_token_for_owner(["esi-corporations.read_blueprints.v1"])
     def _fetch_corporate_blueprints(self, token) -> list:
-        corporation_id = self.corporation.corporation_id
-
         blueprints = esi.client.Corporation.get_corporations_corporation_id_blueprints(
-            corporation_id=corporation_id,
+            corporation_id=self.corporation_strict.corporation_id,
             token=token.valid_access_token(),
         ).results()
         return blueprints
 
     @fetch_token_for_owner(["esi-characters.read_blueprints.v1"])
     def _fetch_personal_blueprints(self, token) -> list:
-        character_id = self.character.character.character_id
-
         blueprints = esi.client.Character.get_characters_character_id_blueprints(
-            character_id=character_id,
+            character_id=self.eve_character_strict.character_id,
             token=token.valid_access_token(),
         ).results()
         return blueprints
 
     @fetch_token_for_owner(["esi-industry.read_corporation_jobs.v1"])
     def _fetch_corporate_industry_jobs(self, token) -> list:
-        corporation_id = self.corporation.corporation_id
-
         jobs = esi.client.Industry.get_corporations_corporation_id_industry_jobs(
-            corporation_id=corporation_id,
+            corporation_id=self.corporation_strict.corporation_id,
             token=token.valid_access_token(),
         ).results()
         return jobs
 
     @fetch_token_for_owner(["esi-industry.read_character_jobs.v1"])
     def _fetch_personal_industry_jobs(self, token) -> list:
-        character_id = self.character.character.character_id
-
         jobs = esi.client.Industry.get_characters_character_id_industry_jobs(
-            character_id=character_id,
+            character_id=self.eve_character_strict.character_id,
             token=token.valid_access_token(),
         ).results()
         return jobs
 
-    def token(self, scopes=None) -> Tuple[Token, int]:
-        """returns a valid Token for the owner"""
-        token = None
-        error = None
-        add_prefix = self._logger_prefix()
-
-        # abort if character is not configured
-        if self.character is None:
-            logger.error(add_prefix("No character configured to sync"))
-            error = self.ERROR_NO_CHARACTER
-
-        # abort if character does not have sufficient permissions
-        elif self.corporation and not self.character.user.has_perm(
-            "blueprints.add_corporate_blueprint_owner"
-        ):
-            logger.error(
-                add_prefix(
-                    "This character does not have sufficient permission to sync corporations"
-                )
+    def valid_token(self, scopes) -> Token:
+        """Return a valid token for the owner or raise exception."""
+        token = (
+            Token.objects.filter(
+                user=self.character.user,
+                character_id=self.eve_character_strict.character_id,
             )
-            error = self.ERROR_INSUFFICIENT_PERMISSIONS
+            .require_scopes(scopes)
+            .require_valid()
+            .first()
+        )
+        if not token:
+            raise TokenError(f"{self}: No valid token found with sufficient scopes")
 
-        # abort if character does not have sufficient permissions
-        elif not self.character.user.has_perm(
-            "blueprints.add_personal_blueprint_owner"
-        ):
-            logger.error(
-                add_prefix(
-                    "This character does not have sufficient permission to sync personal blueprints"
-                )
-            )
-            error = self.ERROR_INSUFFICIENT_PERMISSIONS
-
-        else:
-            try:
-                # get token
-                token = (
-                    Token.objects.filter(
-                        user=self.character.user,
-                        character_id=self.character.character.character_id,
-                    )
-                    .require_scopes(scopes)
-                    .require_valid()
-                    .first()
-                )
-            except TokenInvalidError:
-                logger.error(add_prefix("Invalid token for fetching blueprints"))
-                error = self.ERROR_TOKEN_INVALID
-            except TokenExpiredError:
-                logger.error(add_prefix("Token expired for fetching blueprints"))
-                error = self.ERROR_TOKEN_EXPIRED
-            else:
-                if not token:
-                    logger.error(add_prefix("No token found with sufficient scopes"))
-                    error = self.ERROR_TOKEN_INVALID
-
-        return token, error
-
-    def _fetch_location(self, location_id, token) -> "Location":
-        return Location.objects.get_or_create_esi_async(id=location_id, token=token)[0]
-
-    def _logger_prefix(self):
-        """returns standard logger prefix function"""
-        if self.corporation:
-            return make_logger_prefix(self.corporation.corporation_ticker)
-        return make_logger_prefix(self.character.character.character_name)
+        return token
 
 
 class Blueprint(models.Model):
+    class LocationFlag(models.TextChoices):
+        ASSET_SAFETY = "AssetSafety", _("Asset Safety")
+        AUTO_FIT = "AutoFit", _("Auto Fit")
+        BONUS = "Bonus", _("Bonus")
+        BOOSTER = "Booster", _("Booster")
+        BOOSTER_BAY = "BoosterBay", _("Booster Hold")
+        CAPSULE = "Capsule", _("Capsule")
+        CARGO = "Cargo", _("Cargo")
+        CORP_DELIVERIES = "CorpDeliveries", _("Corp Deliveries")
+        CORP_S_A_G_1 = "CorpSAG1", _("Corp Security Access Group 1")
+        CORP_S_A_G_2 = "CorpSAG2", _("Corp Security Access Group 2")
+        CORP_S_A_G_3 = "CorpSAG3", _("Corp Security Access Group 3")
+        CORP_S_A_G_4 = "CorpSAG4", _("Corp Security Access Group 4")
+        CORP_S_A_G_5 = "CorpSAG5", _("Corp Security Access Group 5")
+        CORP_S_A_G_6 = "CorpSAG6", _("Corp Security Access Group 6")
+        CORP_S_A_G_7 = "CorpSAG7", _("Corp Security Access Group 7")
+        CRATE_LOOT = "CrateLoot", _("Crate Loot")
+        DELIVERIES = "Deliveries", _("Deliveries")
+        DRONE_BAY = "DroneBay", _("Drone Bay")
+        DUST_BATTLE = "DustBattle", _("Dust Battle")
+        DUST_DATABANK = "DustDatabank", _("Dust Databank")
+        FIGHTER_BAY = "FighterBay", _("Fighter Bay")
+        FIGHTER_TUBE_0 = "FighterTube0", _("Fighter Tube 0")
+        FIGHTER_TUBE_1 = "FighterTube1", _("Fighter Tube 1")
+        FIGHTER_TUBE_2 = "FighterTube2", _("Fighter Tube 2")
+        FIGHTER_TUBE_3 = "FighterTube3", _("Fighter Tube 3")
+        FIGHTER_TUBE_4 = "FighterTube4", _("Fighter Tube 4")
+        FLEET_HANGAR = "FleetHangar", _("Fleet Hangar")
+        FRIGATE_ESCAPE_BAY = "FrigateEscapeBay", _("Frigate escape bay Hangar")
+        HANGAR = "Hangar", _("Hangar")
+        HANGAR_ALL = "HangarAll", _("Hangar All")
+        HI_SLOT_0 = "HiSlot0", _("High power slot 1")
+        HI_SLOT_1 = "HiSlot1", _("High power slot 2")
+        HI_SLOT_2 = "HiSlot2", _("High power slot 3")
+        HI_SLOT_3 = "HiSlot3", _("High power slot 4")
+        HI_SLOT_4 = "HiSlot4", _("High power slot 5")
+        HI_SLOT_5 = "HiSlot5", _("High power slot 6")
+        HI_SLOT_6 = "HiSlot6", _("High power slot 7")
+        HI_SLOT_7 = "HiSlot7", _("High power slot 8")
+        HIDDEN_MODIFIERS = "HiddenModifiers", _("Hidden Modifiers")
+        IMPLANT = "Implant", _("Implant")
+        IMPOUNDED = "Impounded", _("Impounded")
+        JUNKYARD_REPROCESSED = "JunkyardReprocessed", _(
+            "This item was put into a junkyard through reprocessing."
+        )
+        JUNKYARD_TRASHED = "JunkyardTrashed", _(
+            "This item was put into a junkyard through being trashed by its owner."
+        )
+        LO_SLOT_0 = "LoSlot0", _("Low power slot 1")
+        LO_SLOT_1 = "LoSlot1", _("Low power slot 2")
+        LO_SLOT_2 = "LoSlot2", _("Low power slot 3")
+        LO_SLOT_3 = "LoSlot3", _("Low power slot 4")
+        LO_SLOT_4 = "LoSlot4", _("Low power slot 5")
+        LO_SLOT_5 = "LoSlot5", _("Low power slot 6")
+        LO_SLOT_6 = "LoSlot6", _("Low power slot 7")
+        LO_SLOT_7 = "LoSlot7", _("Low power slot 8")
+        LOCKED = "Locked", _("Locked item, can not be moved unless unlocked")
+        MED_SLOT_0 = "MedSlot0", _("Medium power slot 1")
+        MED_SLOT_1 = "MedSlot1", _("Medium power slot 2")
+        MED_SLOT_2 = "MedSlot2", _("Medium power slot 3")
+        MED_SLOT_3 = "MedSlot3", _("Medium power slot 4")
+        MED_SLOT_4 = "MedSlot4", _("Medium power slot 5")
+        MED_SLOT_5 = "MedSlot5", _("Medium power slot 6")
+        MED_SLOT_6 = "MedSlot6", _("Medium power slot 7")
+        MED_SLOT_7 = "MedSlot7", _("Medium power slot 8")
+        OFFICE_FOLDER = "OfficeFolder", _("Office Folder")
+        PILOT = "Pilot", _("Pilot")
+        PLANET_SURFACE = "PlanetSurface", _("Planet Surface")
+        QUAFE_BAY = "QuafeBay", _("Quafe Bay")
+        QUANTUM_CORE_ROOM = "QuantumCoreRoom", _("Quantum Core Room")
+        REWARD = "Reward", _("Reward")
+        RIG_SLOT_0 = "RigSlot0", _("Rig power slot 1")
+        RIG_SLOT_1 = "RigSlot1", _("Rig power slot 2")
+        RIG_SLOT_2 = "RigSlot2", _("Rig power slot 3")
+        RIG_SLOT_3 = "RigSlot3", _("Rig power slot 4")
+        RIG_SLOT_4 = "RigSlot4", _("Rig power slot 5")
+        RIG_SLOT_5 = "RigSlot5", _("Rig power slot 6")
+        RIG_SLOT_6 = "RigSlot6", _("Rig power slot 7")
+        RIG_SLOT_7 = "RigSlot7", _("Rig power slot 8")
+        SECONDARY_STORAGE = "SecondaryStorage", _("Secondary Storage")
+        SERVICE_SLOT_0 = "ServiceSlot0", _("Service Slot 0")
+        SERVICE_SLOT_1 = "ServiceSlot1", _("Service Slot 1")
+        SERVICE_SLOT_2 = "ServiceSlot2", _("Service Slot 2")
+        SERVICE_SLOT_3 = "ServiceSlot3", _("Service Slot 3")
+        SERVICE_SLOT_4 = "ServiceSlot4", _("Service Slot 4")
+        SERVICE_SLOT_5 = "ServiceSlot5", _("Service Slot 5")
+        SERVICE_SLOT_6 = "ServiceSlot6", _("Service Slot 6")
+        SERVICE_SLOT_7 = "ServiceSlot7", _("Service Slot 7")
+        SHIP_HANGAR = "ShipHangar", _("Ship Hangar")
+        SHIP_OFFLINE = "ShipOffline", _("Ship Offline")
+        SKILL = "Skill", _("Skill")
+        SKILL_IN_TRAINING = "SkillInTraining", _("Skill In Training")
+        SPECIALIZED_AMMO_HOLD = "SpecializedAmmoHold", _("Specialized Ammo Hold")
+        SPECIALIZED_COMMAND_CENTER_HOLD = "SpecializedCommandCenterHold", _(
+            "Specialized Command Center Hold"
+        )
+        SPECIALIZED_FUEL_BAY = "SpecializedFuelBay", _("Specialized Fuel Bay")
+        SPECIALIZED_GAS_HOLD = "SpecializedGasHold", _("Specialized Gas Hold")
+        SPECIALIZED_INDUSTRIAL_SHIP_HOLD = "SpecializedIndustrialShipHold", _(
+            "Specialized Industrial Ship Hold"
+        )
+        SPECIALIZED_LARGE_SHIP_HOLD = "SpecializedLargeShipHold", _(
+            "Specialized Large Ship Hold"
+        )
+        SPECIALIZED_MATERIAL_BAY = "SpecializedMaterialBay", _(
+            "Specialized Material Bay"
+        )
+        SPECIALIZED_MEDIUM_SHIP_HOLD = "SpecializedMediumShipHold", _(
+            "Specialized Medium Ship Hold"
+        )
+        SPECIALIZED_MINERAL_HOLD = "SpecializedMineralHold", _(
+            "Specialized Mineral Hold"
+        )
+        SPECIALIZED_ORE_HOLD = "SpecializedOreHold", _("Specialized Ore Hold")
+        SPECIALIZED_PLANETARY_COMMODITIES_HOLD = (
+            "SpecializedPlanetaryCommoditiesHold",
+            _("Specialized Planetary Commodities Hold"),
+        )
+        SPECIALIZED_SALVAGE_HOLD = "SpecializedSalvageHold", _(
+            "Specialized Salvage Hold"
+        )
+        SPECIALIZED_SHIP_HOLD = "SpecializedShipHold", _("Specialized Ship Hold")
+        SPECIALIZED_SMALL_SHIP_HOLD = "SpecializedSmallShipHold", _(
+            "Specialized Small Ship Hold"
+        )
+        STRUCTURE_ACTIVE = "StructureActive", _("Structure Active")
+        STRUCTURE_FUEL = "StructureFuel", _("Structure Fuel")
+        STRUCTURE_INACTIVE = "StructureInactive", _("Structure Inactive")
+        STRUCTURE_OFFLINE = "StructureOffline", _("Structure Offline")
+        SUB_SYSTEM_BAY = "SubSystemBay", _("Sub System Bay")
+        SUB_SYSTEM_SLOT_0 = "SubSystemSlot0", _("Sub System Slot 0")
+        SUB_SYSTEM_SLOT_1 = "SubSystemSlot1", _("Sub System Slot 1")
+        SUB_SYSTEM_SLOT_2 = "SubSystemSlot2", _("Sub System Slot 2")
+        SUB_SYSTEM_SLOT_3 = "SubSystemSlot3", _("Sub System Slot 3")
+        SUB_SYSTEM_SLOT_4 = "SubSystemSlot4", _("Sub System Slot 4")
+        SUB_SYSTEM_SLOT_5 = "SubSystemSlot5", _("Sub System Slot 5")
+        SUB_SYSTEM_SLOT_6 = "SubSystemSlot6", _("Sub System Slot 6")
+        SUB_SYSTEM_SLOT_7 = "SubSystemSlot7", _("Sub System Slot 7")
+        UNLOCKED = "Unlocked", _("Unlocked item, can be moved")
+        WALLET = "Wallet", _("Wallet")
+        WARDROBE = "Wardrobe", _("Wardrobe")
+        UNDEFINED = "Undefined", _("undefined")
+
+        @classmethod
+        def from_esi_data(cls, data: str) -> "Blueprint.LocationFlag":
+            try:
+                return cls(data)
+            except ValueError:
+                return cls.UNDEFINED
+
     item_id = models.PositiveBigIntegerField(
         primary_key=True, help_text="The EVE Item ID of the blueprint"
     )
     owner = models.ForeignKey(
         Owner,
         on_delete=models.CASCADE,
+        related_name="blueprints",
         help_text="Corporation that owns the blueprint",
-        # TODO: With next model change: Add related name
     )
     eve_type = models.ForeignKey(
-        EveType,
-        on_delete=models.CASCADE,
-        help_text="Blueprint type"
-        # TODO: with next model change: remove related name
+        EveType, on_delete=models.CASCADE, related_name="+", help_text="Blueprint type"
     )
     location = models.ForeignKey(
-        "Location", on_delete=models.CASCADE, help_text="Blueprint location"
+        "Location",
+        on_delete=models.CASCADE,
+        related_name="blueprints",
+        help_text="Blueprint location",
     )
-    _location_flag_choices = []
-    for choice in EVE_LOCATION_FLAGS:
-        _location_flag_choices.append((choice, choice))
-
     location_flag = models.CharField(
         help_text="Additional location information",
-        choices=_location_flag_choices,
+        choices=LocationFlag.choices,
         max_length=36,
     )
     quantity = models.PositiveIntegerField(help_text="Number of blueprints", default=1)
@@ -464,6 +534,10 @@ class Blueprint(models.Model):
     @property
     def is_original(self):
         return not self.runs
+
+    @property
+    def location_flag_obj(self) -> "Blueprint.LocationFlag":
+        return self.LocationFlag(self.location_flag)
 
     class Meta:
         default_permissions = ()
@@ -541,7 +615,6 @@ class Location(models.Model):
     _SOLAR_SYSTEM_ID_END = 33_000_000
     _STATION_ID_START = 60_000_000
     _STATION_ID_END = 64_000_000
-    _STRUCTURE_ID_START = 1_000_000_000_000
 
     id = models.PositiveBigIntegerField(
         primary_key=True,
@@ -626,10 +699,6 @@ class Location(models.Model):
     def is_station(self) -> bool:
         return self.is_station_id(self.id)
 
-    @property
-    def is_structure(self) -> bool:
-        return self.is_structure_id(self.id)
-
     @classmethod
     def is_solar_system_id(cls, location_id: int) -> bool:
         return cls._SOLAR_SYSTEM_ID_START <= location_id <= cls._SOLAR_SYSTEM_ID_END
@@ -637,10 +706,6 @@ class Location(models.Model):
     @classmethod
     def is_station_id(cls, location_id: int) -> bool:
         return cls._STATION_ID_START <= location_id <= cls._STATION_ID_END
-
-    @classmethod
-    def is_structure_id(cls, location_id: int) -> bool:
-        return location_id >= cls._STRUCTURE_ID_START
 
     def full_qualified_name(self) -> str:
         """Return the full qualified name of this location."""